pyproject.toml
[tool.poetry]
name = "nicehash_sync"
version = "0.1.0"
description = ""
authors = ["Casey Mau <mauza11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
